Finish refactoring notes
NoteEvents are now used instead of Notes, although they have been
renamed to Notes since the original Note class is no longer in use. I'll
refer to them as Notes from here on.

Notes are paired, so one part of the pair is a "note on" event and the
other is a "note off", just like MIDI. These pairs are tracked in code,
allowing things like shifts and duration changes to be done very easily.
This provides convenience while making it easier to translate to MIDI
than the old Notes.

I have also added a Song class for managing all the notes in a song.
This maintains a sorted order for Notes and helps deal with their paired
nature when doing operations on them in the context of the song.

The main motivation for this refactor was to fix a critical issue where
notes were only loaded once by the playback thread and couldn't be
changed. Now, the playback thread reads from the global Song, rechecking
for new Note information every unit. By using the event model, things
like duration changes are respected, and nothing along the lines of
tracking active notes is needed.

Itemized list of changes in no particular order:

- `a` keybind, similar to vim
- `Esc` deselects notes in normal mode
- "Common" note names are used when no scale is given; these are
  determined by whether it's more common for the note to be sharp or
  flat
- MESSAGE is now global, allowing for easier logging
- Some note name-related refactoring to reduce unnecessary Note
  instantiation
- Measure numbers are now 1-indexed in the UI
- Playhead only shifts the view in normal mode
- Add a wrapper to handle KeyboardInterrupts outside of main, allowing
  for graceful shutdown if an interrupt or error occurs during playback
- Some normal mode commands are explicitly blocked in insert mode now