Optimize note rendering to only draw visible notes
Known bug: if the start and end points of a note are out of view, the
note will not be drawn. The only fix I can think of for this is to look
back and ahead by a fixed amount to find notes that start slightly out
of view. It's rare for this to be necessary, and varies by song, so a
comand line argument would probably be the best solution.

Includes some minor Note refactoring for more easily accessing the
correct pair.