Implement Tracks for managing instruments/channels
This is a major refactoring that introduces a Track class instead of
storing the channel and instrument with each note. As discovered during
the implementation of instruments, each channel may only use one
instrument at a time, and therefore storing both can lead to
inconsistency (or alternatively, a ton of program change events).
Furthermore, with each note tracking its channel and instrument
individually, changing the instrument for every note in a channel would
be an expensive operation. Finally, the old method of pre-initializing
every track in FluidSynth to a corresponding instrument was pretty
clunky and did not accommodate multiple channels using the same
instrument well.

To solve all this, a Track class now stores the channel and instrument
for a given track. Each Note's Track is stored by reference, making it
possible to easily retrieve and update this data without needing to go
through the Song. Tracks themselves do little more than provide a nice
data wrapper with some helper methods.

With this implementation, Tracks and channels have a one-to-one
correspondence. When importing, any notes on the same channel in
different tracks will be merged into the same Track, and exported as
such. Theoretically, this should have no effect on the sound of the
song. Notes on different channels with the same instrument are preserved
as expected.

This commit also implements the ability to change the instrument for an
entire channel, as mentioned above. Keybinds to switch between channels
are now available as well.