Refactor main loop and input handling
Instead of one massive function for the main render/input loop, a class
is now used. This allows more granular methods to be extracted without
requiring tons of arguments to access and modify state data.

Input handling has been refactored so that all messages are mapped to
Actions (a new enum) via a dictionary. This removes any dependence on
specific keys from the main input handling code itself, and allows keys
to be more easily rebound and new keys to be added. It also enables a
list of keybindings to be automatically generated, which can be done by
passing the --keymap flag to MusiCLI.