Detect channel volume, track notes during playback
Control changes for control 7 are now interpreted as volume changes (in
accordance with the General MIDI standard). These changes are only
processed once per channel for now. They are also exported along with
the rest of the MIDI.

During playback, active notes are now tracked and stopped when the
playback is paused or restarted. Previously, instruments that didn't
decay were very annoying to work with.