Fix backwards navigation not jumping to chords
get_previous_index() was short-circuiting in most cases because it would compare to the result of bisect_left (which is the note AFTER the previous note) and exit because that note's time would be greater than the time given (which is to be expected). Now the subtraction by 1 BEFORE this sanity check, fixing the off-by-one error.